refactor: ランキング部分のUsecase層を分離
